+++ catalog/docserver/d2.go
@@ -62,6 +62,7 @@ func (ds *DocsServer) exportD2() string {
 }
 
 func (ds *DocsServer) serveD2View(w http.ResponseWriter, r *http.Request) {
+	r = ds.applyCSP(w, r)
 	ds.renderComponent(w, r, D2Page(ds.config.ServiceName, ds.config.DocsPath, ds.exportD2()))
 }